requirements.txt
# Core data manipulationpandas==2.2.3# HTTP requests for Open Electricity APIrequests==2.32.3# MQTT client (v2+ — requires callback_api_version)paho-mqtt==2.1.0# Dashboard and web app (Dash + Plotly)dash==2.18.1plotly==5.24.1# Progress bartqdm==4.66.5# AEMO market data (fallback – not used any more)# nemosis==2.2.0   <-- comment out, we now use OpenElectricity for market